requirements.txt
fastapi
httpx
python-dotenv
python-multipart
uvicorn